safe asum
Added a safe asum wrapper and refactored the safe directory to support more wrappers.
